requirements.txt
django==4.0.6
django-filter==21.1
djangorestframework==3.13.1
markdown==3.3.6
pillow==9.2.0
python-decouple==3.6
dj-rest-auth==2.2.4
django-allauth==0.51.0
drf-dynamic-fields==0.4.0
djangorestframework-simplejwt==5.2.0
python-telegram-==13.13
redis
psycopg2
requests
aiocron
